feat: apply point in purchase

+++ src/main/java/com/purwafest/purwafest/point/domain/constants/PointConstants.java
@@ -1,6 +1,8 @@
 package com.purwafest.purwafest.point.domain.constants;
 
+import java.math.BigInteger;
+
 public class PointConstants {
-    public static final long REFERRAL_POINT_VALUE = 10000L;
+    public static final BigInteger REFERRAL_POINT_VALUE = BigInteger.valueOf(10000);
     public static final String POINT_SOURCE =  "referral";
 }
